+++ tests/testthat/test-padronizar_numeros.R
@@ -142,3 +142,4 @@ test_that("padroniza corretamente - character", {
     gabarito$padronizado_esperado
   )
 })
+